Added -Wall to compilation flags. Removed limitation in partial_sum() where src and dst arrays must be the same type. This allows, say, summing over a C_INT ga and putting the results into a C_LONG ga. Removed warnings generated by -Wall. Using C_ prefixed GA types instead of MT_ prefixed ones. Additional TRACER statements.
git-svn-id: svn://localhost/subsetter/trunk@28 803cb72c-d64d-0410-ae51-a9422f8e9f96

+++ ConnectivityVariable.C
@@ -72,12 +72,12 @@ void ConnectivityVariable::reindex()
                 gather = new int[count];
 
                 NGA_Access(handle, var_lo, var_hi, &buf, var_ld);
-                for (size_t i=0; i<count; ++i) {
+                for (int i=0; i<count; ++i) {
                     subs[i] = &(buf[i]);
                 }
                 NGA_Gather(mask->get_handle_index(), gather, subs, count);
                 // TODO memcpy instead??
-                for (size_t i=0; i<count; ++i) {
+                for (int i=0; i<count; ++i) {
                     buf[i] = gather[i];
                 }
                 NGA_Release_update(handle, var_lo, var_hi);
